+++ db_migration/alembic/env.py
@@ -17,7 +17,7 @@ config.set_main_option("sqlalchemy.url", os.getenv("DB_URL"))
 
 # add your model's MetaData object here
 # for 'autogenerate' support
-from upload_controller_service.dao.db_models import Base
+from ulc.adapters.outbound.db.orm_models import Base
 
 target_metadata = Base.metadata
 # target_metadata = None
