+++ setup.cfg
@@ -14,9 +14,9 @@
 # limitations under the License.
 
 [metadata]
-name = upload_controller_service
+name = ulc
 version = attr: upload_controller_service.__version__
-description = Upload-Controller-Service - The Upload-Controller (ULC) service manages uploads to a S3 inbox bucket.
+description = The Upload-Controller (ULC) service manages uploads to a S3 inbox bucket.
 url = https://github.com/ghga-de/upload-controller-servicee
 long_description = file: README.md
 long_description_content_type = text/markdown; charset=UTF-8
