+++ tests/integration/test_pubsub.py
@@ -21,8 +21,8 @@ from datetime import datetime, timezone
 from ghga_message_schemas import schemas
 from ghga_service_chassis_lib.utils import exec_with_timeout
 
-from upload_controller_service.domain.upload import confirm_file_upload
-from upload_controller_service.pubsub import (
+from ulc.domain.upload import confirm_file_upload
+from ulc.pubsub import (
     publish_upload_received,
     subscribe_file_registered,
     subscribe_new_study,
